generateSourceMaps and preserveLicenseComments can now be used together.
Removed an extraneous blank line that was injected into files in sourcesContent.  This caused the line mappings to be off by one relative to the original source file.  This is significant for debugging in Visual Studio.

Source mappings for flattened files are now expressed in terms of JS statements using esprima, instead of simple source lines.  Also significant for debugging in Visual Studio.

Removed some files that are generated by tests.